+++ packages/ub/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - fix `global.UB.getDomainInfo is not a function`
 
 ## [5.21.3] - 2021-04-19
 ### Changed